REFACTOR - POST endpoint test case for racecar
